Add button component
The idea is to choose different style on the base of the button type by passing props.

For now is added the style for the navigation button.

Edit the type of the component function and add the component in newly created folder. This could be eventually changed in the future. For now is temporary version.