Refactor the left sidebars
Now they aren't treated as layout components, but as separate components in the components folder. Remove the containers, because they are redundant. Enter the code from the containers directly in each feature page (Collections, Products, Favorites). In Profile the code was just slightly transformed, because the code of this page was already there.

Put the main content of each of these pages in the main CSS file as global CSS variable, because we're using it multiple times.