Setup the app server
Render all the products from the server in the main products page.

For that goal we're using axios, installed with npm axios.

The image and the title for each product are now taken from the server (the other stuffs are still hardcoded).

For now there is no available image for all of the products.