Replace the items and ComponentCard props in GenericPage and add children instead of this to render any passed content
Move the creation of ProductCard instances to the Products component, which will pass them as children to GenericPage.

This decouples GenericPage from knowing about ProductCard or the structure of items. GenericPage no longer depends on or knows about ProductCard or the shape of the items it renders. This makes it reusable for other components and scenarios. In that way a single responsibility principle is achieved.