Update the way of fetching the vrscans products and edit the DataContextType to be type instead of interface
Now the vrscans products will be fetched when we open the products page, not after the initial opening of the app.
The other things which are fetched will be changed in that way, too.
When the updateVrscans method is called, it updates the vrscans state. This is a way of caching
The idea is to avoid the fetching of too much data (or redundant quantity data) when we're opening the app.