Do changes related with the shown things in the Collection page
When the collections are rendered, the CollectionsSidebar component is shown and the title is "Collections".

When the collection items are rendered, the ProductsFilters is shown and the title is "Items in 'the current collection title'"

The goal is the child component to focuses on rendering logic, while the parent to handles the sidebar and title logic.